rpi3-nfs: add libckteec.so to the copy.files rule
libckteec.so is now also needed to be able to run xtest, so copy that as
well from the copy.files.

Note that the system is looking for the library at:
 - /lib64/libckteec.so.0
 - /usr/lib64/libckteec.so.0

So technically they should be stored there. Also, since it's looking for
the file ending with "so.0" only, we have to make sure that it exists. A
simple workaround is to use "find -type l ..." so we also get the
symlinks. This will not show up as a symlink at the NFS server, but
that's not a big deal. Likewise, we still copy to /lib/ and not /lib64
and that seems to be all fine.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>
Tested-by: Joakim Bech <joakim.bech@linaro.org> (3.12.0-rc1)